AI-Resume-analyzer/requirement.txt
1.streamlit
2.langchain
3.pdfplumber
4.ollama
